+++ frontend/src/App.jsx
@@ -11,11 +11,9 @@ import Footer from './components/Footer';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import Tools from './pages/Tools';
+import Categories from './pages/Categories';
 import Contact from './pages/Contact';
 import AboutUs from './pages/AboutUs';
-import TermsOfUse from './components/TermsOfUse';
-import PrivacyPolicy from './components/PrivacyPolicy';
-import FAQ from './components/FAQ';
 
 function App() {
   return (
@@ -26,6 +24,7 @@ function App() {
           <Route path="/" element={<Home />} />
           <Route path="/tools" element={<Tools />} />
           <Route path="/tools/:id" element={<ToolDetails />} />
+          <Route path="/categories" element={<Categories />} />
           <Route path="/booking/:id" element={<Booking />} />
           <Route path="/confirmation" element={<Confirmation />} />
           <Route path="/contact" element={<Contact />} />
@@ -33,9 +32,6 @@ function App() {
           <Route path="/admin" element={<AdminPanel />} />
           <Route path="/login" element={<Login />} />
           <Route path="/signup" element={<Signup />} />
-          <Route path="/terms" element={<TermsOfUse/>}/>
-          <Route path="/privacy" element={<PrivacyPolicy/>}/>
-          <Route path="/faq" element={<FAQ/>}/>
         </Routes>
       </div>
       <ToastContainer position="top-right" autoClose={2000} hideProgressBar={false} newestOnTop={true} closeOnClick rtl={false} pauseOnFocusLoss draggable pauseOnHover theme="light" />
