+++ frontend/src/pages/Home.jsx
@@ -1,7 +1,8 @@
-import { useState } from 'react';
-import HomeSearch from '../components/HomeSearch';
+import { useEffect, useState } from 'react';
+import Search from '../components/Search';
 import ToolGrid from '../components/ToolGrid';
 import TopCategories from '../components/TopCategories';
+import SearchResults from '../components/SearchResults';
 
 function Home() {
   const [searchCriteria, setSearchCriteria] = useState({
@@ -53,12 +54,9 @@ function Home() {
 
   return (
     <main>
-      <HomeSearch onSearch={handleSearch} />
+      <Search onSearch={handleSearch} />
       {searchCriteria.searchText || searchCriteria.category ? ('') : (<TopCategories />)}
-      <div className='flex flex-col items-center mt-5'>
-        <h2 className="text-[26px] sm:text-[32px] lg:text-[48px] p-[40px] text-center">Most popular tools</h2>
-        <ToolGrid searchCriteria={searchCriteria} />
-      </div>
+      <ToolGrid searchCriteria={searchCriteria} />
     </main>
   );
 }
